bugfix: Query::rowCount()'s return type must be long according to the type checking rules.

+++ src/jsp/com/cubrid/plcsql/predefined/sp/SpLib.java
@@ -212,12 +212,12 @@ public class SpLib {
             }
         }
 
-        public int rowCount() {
+        public long rowCount() {
             try {
                 if (!isOpen()) {
                     throw new INVALID_CURSOR();
                 }
-                return rs.getRow();
+                return (long) rs.getRow();
             } catch (SQLException e) {
                 Server.log(e);
                 throw new SQL_ERROR(e.getMessage());
